Adding "/dev/block/by-name" symlinks for platform devices
Currently some sources assume the existence of symlinks:
  /dev/block/bootdevice/by-name/<partition_name>.

For example:
  https://android.googlesource.com/platform/system/extras/+/master/cppreopts/cppreopts.rc

However, this symlink is device-specific and implies there is a
"single boot device", which is not True. This change adds the symlinks:
"/dev/block/by-name/<partition_name>" for "platform" devices. Presumably
there shouldn't be two partitions with the same uevent.partition_name
(e.g., "system"). If so, the creation of symlink just fails and is still
backward-compatible.

Bug: 78613232
Test: m init_tests && \
      adb push $OUT/data/nativetest64/init_tests/init_tests /data/. && \
      adb shell /data/init_tests
Test: adb shell ls /dev/block/by-name

Change-Id: I8519c211751e27c8abc128aa61435a6081f9308c