Adds "/dev/block/by-name" symlinks for "platform" devices
Currently some sources assume the existence of symlinks:
  /dev/block/bootdevice/by-name/<partition_name>.

For example:
  https://android.googlesource.com/platform/system/extras/+/master/cppreopts/cppreopts.rc

However, this symlink is device-specific and implies there is a
"single boot device", which is not a requirement. This change adds
the symlinks: "/dev/block/by-name/<partition_name>" for "platform" devices.
Presumably there shouldn't more than one partitions with the same
uevent.partition_name (e.g., "system"). If so, it shows an ERROR log
while creating the same/existing symlink again.

Bug: 78613232
Test: m init_tests && \
      adb push $OUT/data/nativetest64/init_tests/init_tests /data/. && \
      adb shell /data/init_tests
Test: adb shell ls /dev/block/by-name

Change-Id: Iec920b5a72409b6a2bdbeeb290f0a3acd2046b5d