Adding "/dev/block/{platform, pci, vbd}by-name" symlinks
Currently some sources assume the existence of symlinks:
  /dev/block/bootdevice/by-name/<partition_name>.

For example:
  https://android.googlesource.com/platform/system/extras/+/master/cppreopts/cppreopts.rc

However, this symlink is device-specific and implies there is a
"single boot device", which is not a requirement. This change adds
the symlinks: "/dev/block/<type>/by-name/<partition_name>". Presumably
there shouldn't be more than one partitions with the same uevent.partition_name
(e.g., "system"). If so, it bails out when creating the symlink for the
second partition.

Bug: 78613232
Test: m init_tests && \
      adb push $OUT/data/nativetest64/init_tests/init_tests /data/. && \
      adb shell /data/init_tests
Test: adb shell ls /dev/block/by-name

Change-Id: Iec920b5a72409b6a2bdbeeb290f0a3acd2046b5d