Update patch set 1
Patch Set 1:

(12 comments)

High level thoughts:

This would be better if the old keychord driver was removed first.  Some of this code would be simplified if only the new method is supported and since that's the goal, it is where we should start.

This code seems to aggressively call KeychordEnable() for each first service that has a keychord.  That's unnecessary.  By the time keychord_init() is called, we've already parsed all of the init scripts.  You should first generate the list of keychords from the services, then call KeychordEnable() a single time.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>