Update patch set 2
Patch Set 2:

I think this addresses the first round of comments, except:

> How about gets /dev/block/dm-<N> from fs_mgr_get_verity_device_name(), then update it to fstab_rec->blk_device?

If I'm reading this code correctly, it knows ahead of time via a flag in fstab ("avb") that it needs to look in device mapper. We'd need to add a similar flag for dm-linear. If that's okay, any suggestions for a flag name?

Patch-set: 2
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>