Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> (1 comment)
> 
> I think this looks okay, but I'm planning to do a release of platform-tools soon, so I'm going to sit on this for a bit (snapping the branch to aosp/master Soon, but I want to get a few other patches in first to not have to go through cherry-pick hell)

aw, i was hoping to get this in before the release so that the cts harness wouldn't have to use custom-built versions of adb. oh well, if this gets in, at least we'll be able to use the stock version next year.

thanks for the reviews!

Patch-set: 9
Reviewer: Gerrit User 1066743 <1066743@85c56323-6fa9-3386-8a01-6480fb634889>