Update patch set 2
Patch Set 2:

I think the thread + the bits of thread safety needed are going to be better.  It won't be that complex.  We'd just need to make sure that the thread executes on its own without touching any shared state until it has its set of FDs that are opened and state set up, then calling a thread safe register_epoll_handler() and returning.

Besides, if specific open()/ioctl() calls are liable to run for a while, this still puts their costs in the main loop.  The thread will not.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>