init: keychord add debounce timer
Last value in a keycodes can be a negative number that represents
milliseconds of debounce time.  For example:

    on property:init.svc.pAnIcKeRnEl=running && property:ro.debuggable=1
        write /proc/sysrq-trigger "d"
        write /proc/sysrq-trigger "t"
        write /proc/sysrq-trigger "c"

    service pAnIcKeRnEl /system/bin/echo
        oneshot
        keycodes 116 -3000

represents a diagnostic kernel dump and panic trigger if the KEY_POWER
is held for three seconds.

ObligatoryDisclaimer:  If the kernel's input events or init are in
some manner live or deadlocked, then operation and detection may be
blocked.  Whereas if KEY_POWER duration is hardware coded and NMI to
the BSP complex, a deeper and more expert adjustment, the operation
will route.  If this feature is used as a convenience to help diagnose
the live or deadlocked conditions, weigh this shortcoming in triage of
such serious issues.

Test: manual
Test: init_tests --gtest_filter=keychords.esc_too_short
Test: init_tests --gtest_filter=keychords.leftalt_too_long
Bug: 79553608
Bug: 76217789
Bug: 64114943
Change-Id: I10e9721b94c81bb8f22b32356e4d1ec4e59de006