Update patch set 4
Patch Set 4:

(2 comments)

Bug that triggered this enhancement is moving towards adjusting the bootloader.

The goal of this change is to ensure that one can create this policy in Android and device properties.

IMHO having two ways of doing things is not desireable so we continue with a -2 for this change.

Patch-set: 4
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>