Update patch set 2
Patch Set 2:

The benefit is all about maintenance and cleanup/reduction of code.

At the request of tomcherry@ because keychord is a low likelihood trigger, performance is not an issue. One can directly compare std::set directly, whereas std::vector requires a pair of double-loops to do so. This CL is to demonstrate feasibility and show the maintenance gains. By using it we solve the (academic) problem of multiple identical registrants and eliminate the defunct keychord id from all places in the service structure.

Typically there is _one_ registered keychord, and only on Pixel devices. No partner, AndroidTV, AndroidAuto or Clockwork device uses this feature so moving towards simplicity and maintainability is of some use. With the introduction of the duration/debounce change, we may have maybe two registrants and possible users outside of core Pixel products. Matching chords will still be a low volume occurrence.

Given this possible controversy, now two reasons to move, I will spend a few moments to refactor the change sets to place this _after_ the class refactor of keychords. The other benefit gained in backup plans regarding the Planned-To-Be-Deprecated Android keychord driver in the kernel.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0