Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Now that CL:689777 is merged, put this behind USE_LOGICAL_PARTITIONS build flag?
> 
> Or do you need to work on this for more

The current plan is to implement this outside of fs_mgr, so it can be linked with libavb. So if this CL landed it would be short lived. I don't have an opinion either way whether it's worth it or not.

Patch-set: 5
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Work-in-progress: false