Update patch set 3
Patch Set 3:

(1 comment)

> I did try to separate it out, got _uglier_ and when I do not bundle the timeout in the keycodes for matching, then for example:

Understood.  At the end of the day, I only suggested std::set to see if it makes things cleaner.  If it's not, then the idea can be dropped.  Perhaps having a vector and Keychords::Register() sort the input before emplacing it is a better option.  It would continue that operator== works and that timeouts are at position 0, but skip some of the avoid of the questions here.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>