Update patch set 7
Abandoned

Testing has confirmed this makes llkd 2.5% faster, which although somewhat repeatable results with >40 samples, is well below the noise so not a guarantee sample to sample. llkd takes less than 0.1% of the CPU on a generous sampling rate, so an imperceptible gain overall. The gain in llkd algorithms is maybe less likely to be impacted by a live lock (file open blocking if proc is deadlocked?), but can not be counted on as prevention and may be only an academic or FUD gain. The loss is a higher memory footprint (especially if mlockall) and the risk that if incorporated into another health daemon (lmkd?) will not share the _SC_OPEN_MAX without tuning exhausting the available descriptors.

As a result I am viewing this as too much complexity, risk and maintenance to be worth pursuing.

Patch-set: 7
Status: abandoned
Tag: autogenerated:gerrit:abandon