Build init with shared libraries
recovery partition now supports shared librarys. Therefore, init can now
be built as a dynamic executable both for normal and recovery modes.

To save save in the recovery mode, not all libs are dynamically linked.
Libs that are only used by init in the recovery mode are still
statically linked.

Bug: 63673171
Test: `adb reboot recovery; adb devices` shows the device ID
Change-Id: I30c84ae6a8e507e507ad0bb71acad281316d9e90