bootstat: Add \nBOOTSTAT suffix to value in persist.sys.boot.reason
To ensure a surprise reboot does not take the last boot reason on
face value despite if coming from more than one boot sessions ago.
We keep the value around in the current relevant boot session with
a suffix of \nBOOTSTAT.  If we reboot, we will consider the value
out of date and ignore it.

Test: boot_reason_test.sh
Bug: 86671991
Change-Id: If85750704445088fd62978679ab3a30744c46abb