fs_mgr: add overlayfs handling for squashfs
Add fs_mgr_is_overlayfs() to determine if all the conditions are
correct for allowing overlayfs to be activated at time of call.
Use results to mount overlayfs read-only if all conditions are met.

Conditions are that filesystem is squashfs, read-only, debug build,
kernel supports overlayfs.  A default is available, but an override
is an fstab overlay entry with upperdir=, lowerdir= the same mount
point, workdir= and wait.  The default is a valid /data/overlay/
or /cache/overlay/ directory, .../upper/<mount_point> and
.../work/<mount_point>, associated with each system partition
<mount_point>.

Will set the sepolicy to context as rootcontext= option,
<mount_point>, or u:object_r:<mount_point>_file:s0 in that order.

Test: compile
Bug: 109821005
Change-Id: Ie9b4e9ca0f2bdd9275ac2e2e905431636948e131