Update patch set 3
Patch Set 3:

(9 comments)

Two big changes in this version:

(1) It's now a separate liblp static library within fs_mgr. I did this so the code could be used when host_supported is true, which makes it easier to generate a metadata image as part of the build.

(2) LpMetadataHeader was really clumsy to update, backup, and A/B. There was no deterministic way to figure out its maximum size, if the metadata was somehow corrupt or invalid. This made it harder to locate additional copies. The new approach is to store a very simple descriptor in the first and last sectors of the partition, described by LpMetadataGeometry. This tells us the maximum size of the metadata and how many copies are needed for A/B. The old "backup locator" goes away.

The geometry is never updated. When writing new metadata, it is verified as being compatible with the existing geometry.

Patch-set: 3
Reviewer: Gerrit User 1045980 <1045980@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1121838 <1121838@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Work-in-progress: false