fs_mgr: add overlayfs for small-space filesystems
If free space on the systems partition is less than Android
properties persist.adb.remount.overlayfs.minfree, or
ro.adb.remount.overlayfs.minfree, default 1%, then we may use
overlayfs to override the filesystem.

0 will turn space checking feature off, 100 will force overlayfs on
all filesystem types.

It is possible there is more free space on the filesystem, then is
available in the override upper directory storage.  We do not check
this because the advantage of overlayfs is that the overrides can be
rolled back, by adb enable-verity;  Whereas a traditional remount of
a read-writeable filesystem can not, without reflashing the original
content.

Test: boot
Bug: 109821005
Change-Id: I44f4e1d1955308d5b37024d3c9a3a822a64533b2