Update patch set 7
Patch Set 7: Presubmit-Verified+1

=== Finished presubmit run: L08000000182291514 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/701577?revision=7&workplanId=L08000000182291514




TreeHugger finished with: 21 passed.





WARNING: The tip-of-tree(4839523) for git_master/errorprone may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4839523) for git_master/gce_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4839523) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






WARNING: The tip-of-tree(4839523) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4839523) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4839523) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





This is some super-exciting work! Well done!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 7
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1