Update patch set 2
Patch Set 2:

> Patch Set 1: Verified+1 Presubmit-Ready+1
> I didn't include a unit test for this because kernels don't have dm-zero enabled by default, and that makes it harder to create building blocks for dm-linear.

Do we want to remain functional if dm-zero is disabled? It means sparse regions will be backed by (precious) storage, but at least we have a volume to work with. Android sparse for filesystem and verity metadata marks blocks of zeros, whether part of the filesystem or not, we may have to rely on verity-fip to report the blocks back as zeros.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>