Update patch set 8
Patch Set 8:

(1 comment)

> Patch Set 8:
> 
> (1 comment)
> 
> > 1) tomcherry@ finds the directory selection code complicated, I like that it will _not_ drop existing overlay content if the filesystem gets full, that is a design requirement. I do not believe the code can be simplified unless we make it uglier by adding two scan loops.
> 
> I also do not want to drop the content.  I would rather *always* use /cache or *always* use /data (as in we pick one) unless a user specifically opts into using the other one.  This drops the complexity and maintains that requirement.
> 
> If the mount that is selected gets corrupted (permissions, etc) or gets filled, I'd like that information to present back to the user instead of trying to use a different directory which may have behavior that they don't expect.

Ok, so for Pick One, we will need two loops. One pass to find the one that already has override _content_ and select that one regardless if it is r/w or r/o to provide the "pick-one". A second loop to search for the first r/w combination. Do you disagree fundamentally with me merging the two loops, we search for both conditions? Or what is the complication you are concerned about?

(NB: as discussed in the past, we can deal with the slippery slope of how we handle layering overlayfs entries (eg: /cache is full, the also overlays /data) as a future enhancement).

Patch-set: 8
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>