Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> (1 comment)
> 
> > Patch Set 8:
> > 
> > (1 comment)
> > 
> > > 1) tomcherry@ finds the directory selection code complicated, I like that it will _not_ drop existing overlay content if the filesystem gets full, that is a design requirement. I do not believe the code can be simplified unless we make it uglier by adding two scan loops.
> > 
> > I also do not want to drop the content.  I would rather *always* use /cache or *always* use /data (as in we pick one) unless a user specifically opts into using the other one.  This drops the complexity and maintains that requirement.
> > 
> > If the mount that is selected gets corrupted (permissions, etc) or gets filled, I'd like that information to present back to the user instead of trying to use a different directory which may have behavior that they don't expect.
> 
> Ok, so for Pick One, we will need two loops. One pass to find the one that already has override _content_ and select that one regardless if it is r/w or r/o to provide the "pick-one". A second loop to search for the first r/w combination. Do you disagree fundamentally with me merging the two loops, we search for both conditions? Or what is the complication you are concerned about?
> 
> (NB: as discussed in the past, we can deal with the slippery slope of how we handle layering overlayfs entries (eg: /cache is full, the also overlays /data) as a future enhancement).

I wouldn't check for content.  I'd check that only one exists and fail to mount anything if there are multiple mount points prepared.  I'd then make the setup code ensure that it only creates one mount point, and itself fail to setup the other mount point (which the user would have to override to even request this) if the other mount point exists.

We shouldn't overlay the entries.  A device should have a default that it uses.  Users should be able to override these when preparing the mount points, but do so full-well knowing the caveats, e.g. /cache is likely to mount earlier but be smaller.

Patch-set: 8
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>