Update patch set 8
Patch Set 8:

> IMHO multiple directories setup _blank_ should not be considered an error. The error should be if multiple entries have override data in the upper directory.

Why not?  Is there a valid use case for multiple blank set up directories?  If not, then we should disallow it and gain the advantage of simplifying the code that determines which directory to use.

We don't need a property or a bringup engineer's decision if we do the above.  We end up with a few simple scenarios:

1) LRAP with /scratch -> overlayfs mount happens in first stage mount if /scratch is set up for it, /data and /cache are never used.

Non-LRAP:
2) User does adb disable-verity -> prepare /data for overlayfs mounts and on subsequent boot they are used
3) User does adb disable-verity and explicitly specifies /cache (we'd need to give them an option here) -> prepare /cache for overlayfs mounts and on subsequent boot they are used
4) 3) happens after 2) happens -> error to the user they they're already using /data and give them instructions on how to tear down the /data mount.

What cases do those not cover?

Patch-set: 8
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>