Update patch set 8
Patch Set 8:

It feels like I am repeating the design document over and over again... I will state my case with fragments from it rather than short paraphrases as I have done so far ...

Basically I disagree with failure to mount an existing override. I disagree with failure to mount if two candidate directories within _one_ mount phase appear both to be setup, but upper has yet to be populated. These failures are _unnecessary_ and represent a degraded product that does not serve the _customer_ of this feature.

[TL;DR]

> Patch Set 8:
> 
> > IMHO multiple directories setup _blank_ should not be considered an error. The error should be if multiple entries have override data in the upper directory.
> 
> Why not?  Is there a valid use case for multiple blank set up directories?  If not, then we should disallow it and gain the advantage of simplifying the code that determines which directory to use.
> 
> We don't need a property or a bringup engineer's decision if we do the above.  We end up with a few simple scenarios:
> 
> 1) LRAP with /scratch -> overlayfs mount happens in first stage mount if /scratch is set up for it, /data and /cache are never used.
> 
> Non-LRAP:
> 2) User does adb disable-verity -> prepare /data for overlayfs mounts and on subsequent boot they are used
> 3) User does adb disable-verity and explicitly specifies /cache (we'd need to give them an option here) -> prepare /cache for overlayfs mounts and on subsequent boot they are used

I agree there should be a way to override, and we have presented the API. However, adb disable-verity <directory> is non-intuitive and needs education, and a user specified backing for overlayfs _must_ be of the set hard-coded in the library to support init first stage mount. 

I disagree that it _needs_ to be overridden for any of the scenarios, save for suboptimal solutions with an understanding of the caveats. Boot time always provides the optimal solution with the least caveats without intervention. The override is for a less optimal solution, for example wanting to select later mount /data to allow for near unlimited space. This CL should represent pure and optimal automation, with no overrides, a later CL will add the overrides.

Early mount /data coincident with /cache is the only real automatic selection _I_ want BTW, works handily on the Hikey and dev-board products, selecting /data in that specific case, and deals handily with this scenarios. All others are pretty clear cut at boot time. I do want some hardening though, for instance placing priority on _populated_ upper, that was not in my design document, but makes sense.

BTW, for another scenario, /cache can mount at init first stage. To fix this is an easy step for the bringup engineer, and should not require them to adjust more than just the fstab.

> 4) 3) happens after 2) happens -> error to the user they they're already using /data and give them instructions on how to tear down the /data mount.

adb overlayfs-backing <directory> or adb disable-verity <directory> whatever, can _move_ content without having to supply an instruction sheet. Leave this for a later CL.

> What cases do those not cover?

The case where we want to keep user or bringup intervention to a minimum ;-}

Always remember that in almost all debug setup scenarios, if it requires thought, or an audit of all efforts by placing content in two places, it is implemented at the lowest priority. If problems arise, the immediate response is to move on and to do a fastboot flashall, not to go and figure out what can be adjusted to fix the problem. We are effectively quibbling over 10 lines of code that would generally _prevent_ the failure scenarios. I feel we do a disservice by dropping 2 lines of code that help survive handily and instead report a failure that no one sees until they, or the system, try to use the overridden content. This will only serve to confuse the developer, he will not see a log message, he will see that his recent cycle of code did not fix a problem, and he will think he needs to do more work on the code. Fsilure, when failure is not required, is a WTF. Any failure scenarios that can automatically work around the problem must be considered to harden to reduce the chances of the debug feature being declared prone to failure or noisily balking and reporting problems.

The customer for this feature is the _developer_, and must be presented fully functional, follow an existing standard of behavior and presented KISS to _them_.

I believe the case where _two_ directories contain overrides is a ripe situation to overlay an overlay (part of my architectural concerns section). I do not agree to fail and leave the developer wondering what happened to his updated code. But that piece of hardening can wait until a later CL. The chances of two directories being presented at boot time at some early mount phase is few _except_ for the dev board products AFAIK. dev boards represent one of our _customers_. Both dirs being populated is even less likely, but it can only happen if the developer changes an override (without prejudice or controlling app like adb overlay-backing <directory> to validate), and that should be another scenario we may cover, where the developer started out with /cache as automatically selected, but then _later_ wanted late-mount /user to continue with because /cache was overflowed. But I feel is outside the scope of this CL and for _them_ to implement ;-} as an enhancement.

Patch-set: 8
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>