Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Hey Yi. I understand this makes things slightly better (since it is the pointer type), but did you uncover any bugs? I tend to prefer an incremental approach here since a change like this is going to make back-fixes more difficult. Most new code shouldn't be using most of these classes anyway as well.

Binder has had problems in the past with undefined behavior, so we really want to turn on as many warnings as we can here. Having better strong-typing for pointer usage in C++ also helps prevent bugs in the future, so just looking at the past isn't the only metric.

Seeing that Tom also commented on this about conversion to C++, is that something that would be possible? I would much prefer if we can do something like that instead too. It's time to remove or convert as much C to C++ as we can.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>