Update patch set 1
Patch Set 1: Presubmit-Verified+1

=== Finished presubmit run: L92700000189659565 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/715750?revision=1&workplanId=L92700000189659565




TreeHugger finished with: 23 passed.





WARNING: The tip-of-tree(4886385) for git_master/cf_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4886385) for git_master/errorprone may be broken.
The presubmit build passes with the last-known-good-build(4886214), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4886385) for git_master/gce_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.










WARNING: The tip-of-tree(4886385) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






WARNING: The tip-of-tree(4886385) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886214), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4886385) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886194), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4886385) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4886206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





This is an awesome step forward. Great job!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 1
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1