Update patch set 16
Patch Set 16:

> Patch Set 16:
> 
> (16 comments)
> 
> A lot of style comments that will significantly help readability/reviewability inline.
> 
> I see in the comments that the plan is to use the classes to rewrite engine.cpp.  Do you have a CL that does that, even if WIP?  Seeing the end usage of the classes will help understand how this all fits together.
> 
> It doesn't seem like there is a good match between the actions in engine.cpp and the Command classes.  If the only user of these classes is the FastBootDriver class, then it would be better if those classes were just coded directly as functions.
> 
> What is the difference between the new FastBootDriver class and the existing functions in protocol.cpp?  These seem to have a lot of overlap and should be merged if any changes are to be done.

Hi,

Thanks for your feedback!

I do kinda have a CL, but I need to fix some thing. The idea will make engine.cpp super clean.
If you notice that it current is just queuing Action structs. Each of these Action structs hold an enum code that corresponds to what is now a "Command", the data needed for that "Command", a message to print before, and a callback function that runs after the "Command". 

Thus, if engine.cpp will just subclass all the Commands to have a virtual Begin() and After() method (for clarity call this class EngineCommand). These Begin() and After() methods will be called before a command is executed and after a command is executed respectively. This will be used to perform the job of the callbacks in the current Action struct. 

Then it is super simple, I just keep a deque<EngineCommand> of the everything that needs to be done. When processing the queue, I pull each command out and do the following:
* cmd.Before();
* fasboot.Command(cmd); // since EngineCommand extends Command
* cmd.After();

The commands to queue a command, for example "fb_queue_erase(partition)", will just become:
EngineEraseCmd cmd(partition);
cmds_deque.push_back(cmd);

^ I hope that makes sense. I will work on a CL to make it clearer.

Also protocol.cpp should be deleted, as FastBootDriver was designed to replace it.

Patch-set: 16
Reviewer: Gerrit User 1336956 <1336956@85c56323-6fa9-3386-8a01-6480fb634889>