Update patch set 18
Patch Set 18:

> Patch Set 16:
> 
> (4 comments)
> 
> Okay, I understand the goal.
> 
> What is the benefit of having the middle layer Command class though if FastbootDriver already has the various commands exposed as functions?
> 
> Instead of having to call three functions (Before(), Command(), After)), you could have a class hierarchy of the higher level operations from engine.cpp (not based off of the Command classes you already have).  These would have a single 'execute' function that takes the a FastbootDriver parameter.  It would do the required 'before' work, call the appropriate command (Flash(), Download(), GetVar(), etc) from FastbootDriver, then do the appropriate 'after' work.
> 
> Does that make sense?  You'd have the same interface but a more simple implementation.
> 
> > Also protocol.cpp should be deleted, as FastBootDriver was designed to replace it.
> 
> Is it deletable as of this CL?  It would be good to actually remove it if FastbootDriver is superseding it.

Really appreciate your quick feedback. I need to get this API stable, because I am working on a fastboot testing program that is using this.

>> Does that make sense?  You'd have the same interface but a more simple implementation.
Yeah, I get exactly what you are saying. The real reason I decided to to it this way is for two reasons. First the stuff that needs to be done by engine in Before() and After() is nothing related at all to fastboot. It is almost entirely just printing stuff to the console or mutating the internal state of engine.
Like for example with a SparseTransfer:
Begin() would print something like: "Writing sparse" and After() would be like "OKAY [  3.624s]"

The other reason is my testing class I mentioned above needs to be able to create "evil" malformed commands. This is a simple as extending the Command base class to do bad things in Execute(). Then I can simply pass this command into fastboot.Command()

>> Is it deletable as of this CL?  It would be good to actually remove it if FastbootDriver is superseding it.
Just deleted it! :)

Patch-set: 18
Reviewer: Gerrit User 1336956 <1336956@85c56323-6fa9-3386-8a01-6480fb634889>