Update patch set 14
Patch Set 14:

IMHO, reading (not writing) incorrect sysprops is closer to a compatibility problem than a security problem. I know that client checking can be avoided, but it still saves a lot of developers with good-will from making mistakes that can break backwards compatibility.

For example, vendor accessing system-owned private sysprop is bad and should be prevented not because it is causing security problems, but because it will cause compatibility problems when the system is updated independently from the vendor. (Of course, in some cases that might also cause security problems, but I am not saying that we shouldn't use on SELinux at all. We can just use it as it is, while doing another compatibility-related checks using these classes)

Patch-set: 14
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>