Update patch set 14
Patch Set 14:

I'll buy that argument.

But, the checks need to be done at compile time.  The tool that generates these APIs should simply not create the functions for the wrong clients.

> For example, we can allow a sysprop to be accessible from platform libraries while disallowing it from vendor libraries even when the libraries are loaded in the same process.

To use your own example, that specifically is something best handled compile time.  No amount of checking UIDs or SELinux is going to let you do that, yet it would be trivial to have the tool simply not generate the APIs for those properties for vendor code.

Really just to summarize my thoughts here: I'm fine with idea of having the manifest of properties and generating type safe property APIs that handle the parsing.  But the code should come with 0 runtime overhead.  Most, if not all, of what you want to accomplish can be done by generating the right headers on the host side and that is the best implementation in my opinion.

You can add all of the necessary parsers / setters to android-base and generate header-only code that essentially has

namespace platform {
  namespace ro {
    namespace build {
      ...
      std::optional<string> date() {
        return GetStringProperty("ro.build.date");
      }
      std::optional<string> description() {
        return GetStringProperty("ro.build.description");
      }
      std::optional<string> fingerprint() {
        return GetStringProperty("ro.build.fingerprint");
      }
      ...
    };
  };
};

On a side note, have you done any analysis into how many users of some of these parsers you'll have?  In particular the lists of integers/doubles?  I literally do not have any of these matches on my walleye.

Patch-set: 14
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>