Update patch set 3
Patch Set 3:

> Patch Set 3: Verified-1
> 
> This breaks completion on mac (which uses bash 3.2):
> 
>   bash-3.2$ source build/envsetup.sh
>   bash: type: compopt: not found
>   bash: type: compopt: not found
>   bash-3.2$ adb <tab>bash: type: _init_completion: not found

https://github.com/mathiasbynens/dotfiles/issues/544
This issue is exactly the same as what we encountered, upgrading bash via brew will resolve this issue. Many built-in commands such as compopt and declare are not available in out-of-box bash in MacOSX.

Perhaps we have to test BASH_VERSION > 4 in envsetup.sh and inform users to upgrade bash to eliminate this issue since 3.x is really outdated.

Patch-set: 3
Reviewer: Gerrit User 1311067 <1311067@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0