Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Andrew, is there any reason we can't just invoke 'modprobe' to load the named module in the chain, and that will do a similar thing and load all of the module's dependencies?
> 

This can be done, and is actually how I originally implemented this. The first thing that needed to be done was to fixup the modprobe shipped with toybox, it currently does not search the right Android paths for modules (vendor/). When doing this I realized how little code would be needed if this was all just done in init. We would also then not depend on an external modprobe and init could even become the systems modprobe tool (like with ueventd and watchdogd).

I'm okay with either if there are strong opinions one way or another.

> I can see one disadvantage is that modprobe is fire-and-forget, there's no recourse if it fails, but I'm not sure what error handling you could usefully do from ueventd/init anyway.

If one attempts to load modules before the filesystem they are on is ready we can do deferred loading, much like how is done with firmware in init/firmware_handler.cpp.

Patch-set: 1
Reviewer: Gerrit User 1339568 <1339568@85c56323-6fa9-3386-8a01-6480fb634889>