Update patch set 2
Patch Set 2:

High level thoughts:

The memory leak fixes with bootimg_utils.cpp and great and well overdue.  That and adding reboot-bootloader to fastboot.cpp can/should go in as stand alone changes.

As for engine.cpp itself, is the queue actually helpful in the first place?  The fastboot app itself is synchronous, so it seems that the only thing we gain from the queue is that we get to do a limited amount of input validation before executing the queue, which enh@ has separately told me he's willing to forgo if it means simplifying this code.  Regardless, this itself is actually broken, since flashall and update call check_requirements(), which flushes the queue:

 $ fastboot flash boot bad_name flashall
fastboot: error: cannot load 'bad_name': No such file or directory
 $ fastboot flashall flash boot bad_name
--------------------------------------------       
Bootloader Version...: 8996-012001-1710040120
Baseband Version.....: 8996-130091-1802061512
Serial Number........: HT6640300170
--------------------------------------------       
Checking product                                   FAILED

Device product is 'sailfish'.
Update requires 'hikey960'.

fastboot: error: requirements not met!

And check_requirements() itself is almost definitely better implemented in a synchronous fashion.  It essentially is currently, since ExecuteQueue() is called at the end of check_requirements(), but since it goes through the queue, instead of having requirements checking isolated to a single area in fastboot.cpp, it gets spread out among 3 files, making this code way more complex than is needed.

That is to say, besides how engine.cpp is legacy code, the actual mechanism looks like it's a relic too and instead of making a better queue it may be better to re-evaluate how the queue is used in the first place.

Lastly, are engine.cpp or the Engine class something that your project is going to use or just FastBootDriver?  If only the latter, I'd focus on the project instead of worrying about cleaning up this code, since there's plenty of clean up to do, but it has a much lower ROI.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>