Update patch set 3
Patch Set 3:

> I don't think the queue mechanism is broken at all since on failure remaining items are not processed.

I mean that its one benefit (up front checking of arguments) is currently broken, so keeping the queue should not be a priority.

> Yeah, I agree the queue is not necessary. The commands could be run immediately after the command line flag is encountered instead of just queuing it. 

I'm glad we agree here.

> 1. It separates out the boilerplate of actually performing the fastboot action and doing all the associated pretty-printing and error handling. 

I'm sure there are other approaches that can accomplish this too.  

Before() itself is essentially boiler plate, since these prints can be positioned at the top of Execute() without penalty.  

Nearly every Execute() command is just a pass through to an existing command in FastBootDriver, so its existence is essentially boiler plate.

After() is somewhat questionable to me too.  Outside of the default usage, nearly all of the code contained in these After() blocks is better done in the caller (assuming synchronous code)

Require -> Isn't needed if check_requirements() is rewritten to be synchronous and to use FastBootDriver::GetVar().

GetVarStore -> Again isn't needed if check_requirements() is synchronous.  Doesn't even seem to be needed now, since we have a synchronous GetVar() already.

Display -> Only used to print the contents of a variable, which is trivially done with FastBootDriver::GetVar() if everything is synchronous

Reboot -> It's easy to just not do anything in the caller

Notice -> This literally just is used to print a string to console, which can be totally removed if everything is handed synchronously

In fact all of 5 of those classes can be removed if the queue is removed, which seems to be a much bigger reduction in unnecessary code.

> 4. It completely isolates performing commands (engine.cpp), with figuring out what commands need to be run (fastboot.cpp).

You already have this separation with your last CL.  FastBootDriver contains the commands and fastboot.cpp figures out what commands should be run.  The only thing that prevents this from being clearer is the abstraction of engine.cpp which muddies the water.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>